+++ feature/signin/src/main/java/team/retum/signin/navigation/SignInNavigation.kt
@@ -1,8 +1,9 @@
-package team.retum.signin
+package team.retum.signin.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
+import team.retum.signin.ui.SignIn
 
 const val NAVIGATION_SIGN_IN = "signIn"
